backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.100.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
distro==1.9.0
docstring_parser==0.18.0
filelock==3.29.0
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2026.4.0
gunicorn==26.0.0
h11==0.16.0
hf-xet==1.5.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.14.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
markdown-it-py==4.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.2
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rich==15.0.0
shellingham==1.5.4
sniffio==1.3.1
tokenizers==0.23.1
tqdm==4.67.3
typer==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
Werkzeug==3.1.8
